backend/pyproject.toml
[project]
name = "gym-app-backend"
version = "1.0.0"
description = "Gym App Backend API built with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy==2.0.36",
    "aiosqlite==0.20.0",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "email-validator==2.2.0",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.2.1",
    "python-multipart==0.0.18",
    "python-dotenv==1.0.1",
    "greenlet>=3.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.1",
    "ruff==0.8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
